requirements.txt
requests
beautifulsoup4
PyQt5
pygame
lxml
pillow